Add gameplay logic to move in key with game mode
Gamemode selection passed through prop and handled
via switch case directly to click handler.

The API as written can detect the game board and
determine the current player without need for it
to be passed through. As such, a switch statement
is the most readable solution, however blase it may
seem.

Axios requests handle calls asynchronously and work
to unpackage the updated API, which