Create top-level navbar
The Navbar is set to animate in on page load,
with the button and navbar added to the root layout
in order to be present on every page.

The animation is implemented using Framer Motion, with
liberal use of variants and smoothing.

TODOs also added to note future additions that I want
to implement. These are largely aesthetic, and I need
to focus on content for a while, so they'll be added
more slowly.