Preview (#1)
* Reroute tictactoe to separate path location

Create placeholder for index page

* Create template title in root layout

* Add Navbar component and reformat DarkModeBtn

* Update border behavior for light/dark mode

* Add Navbar styling and iterate mobile navbar disp

This commit is a intermediate update for the
switching of workstations, and does not represent
a full update in functionality.

* Update custom tailwind color palette in 'standard'

* Complete standard color palette

* Install react-icons

* Tighten animation for DarkModeBtn

* Refine animations for DarkModeBtn

* Create navbar skeleton and animation presets

* Remove unnecessary viewport check causing errors

* Add and configure onTap for accessibility

* Create top-level navbar

The Navbar is set to animate in on page load,
with the button and navbar added to the root layout
in order to be present on every page.

The animation is implemented using Framer Motion, with
liberal use of variants and smoothing.

TODOs also added to note future additions that I want
to implement. These are largely aesthetic, and I need
to focus on content for a while, so they'll be added
more slowly.

* Reconfigure DarkModeBtn as motion.div

* Refine index page with basic palette and lorem

* Add TODO to navbar.tsx

* Fini

* Add null transition elements to Nav and add TODO

Typescript error in navbar.tsx is being thrown
erroneously, want to resolve it without damaging
future typing for Framer Motion animation variants

* Rollback null values in motion transitions

Prior resolution of TODO started throwing TypeScript
errors for motion variant typing. I falsely believed
this would not interfere with build integrity, but
npm run build threw an error and returned 1. Going
to investigate further to implement smoothing while
preventing this error.

* Adjust navbar timing to better present on mobile

* Removed onTap listener to fix mobile behavior

* Refactor flexbox/tailwind for better display

* Add custom hook to obtain window height

Next.js's use of SSR makes the use of
window.innerWidth extremely difficult. The
use of this hook is meant to circumvent that
and allow npm build to properly exit without
the 'window is undefined' error

* Rework navbar and toggle behavior for mobile

* Fix typo and slightly adjust animation timing

* Adjust timing and reduce opacity for ultrawides

* Resize nav toggle to conserve viewport space

* Add clarification in TODO

* Begin MDX configuration

This commit was made mid-setup to facilitate
switching between workstations. Don't expect any
real changes here.

* Update NavToggle animation

* Fix typo in custom hooks

* Update animation on NavToggle

* Add Link to navbar and detect current route

* Add github button to navbar

* Install contentlayer and fix npm vulnerabilities

* Install rehype-slug and create contentlayer config

* Add .contentlayer generations to .gitignore

* Update tsconfig paths

* Add profile picture to navbar

* Add animation to profile picture in navbar

* Reconfigure flexbox in tictactoe for better sizing

* Begin working on homepage

This is an intermediate commit meant to facilitate
changing workstations

* Update flex properties for navbar

* Update scaling on tictactoe

* Create temp landing page

* Add temp landing page preview